move app types into types file
